+++ src/tools/registry.ts
@@ -1,10 +1,6 @@
 import { ITool } from '../types/common.js';
 import { Logger } from '../services/logger.js';
 
-import { ReadFileTool } from './file-operations/read-file-adapter.js';
-import { ListDirectoryTool } from './file-operations/list-directory-adapter.js';
-import { GetFileMetadataTool } from './file-operations/get-file-metadata-adapter.js';
-
 import { 
   WorkspaceSetupTool
 } from './project-management/index.js';
@@ -45,14 +41,6 @@ export class ToolRegistry {
    * Register default tools
    */
   private registerDefaultTools(): void {
-    // 基本ツール
-    
-    // ファイル操作ツール
-    this.registerTool(new ReadFileTool());
-    this.registerTool(new ListDirectoryTool());
-    this.registerTool(new GetFileMetadataTool());
-
-    
     // プロジェクト管理ツール
     this.registerTool(new WorkspaceSetupTool());